김예영 : OpeningLevel Timer기능 및 점프 기능 업데이트

+++ APIApp/GameEngineCore/GameEngineLevel.cpp
@@ -251,14 +251,14 @@ void GameEngineLevel::ActorsRender(float _DeltaTime)
 	}
 
 
-	//Text 출력
+	//출력
 	{
 		TextOutStart = float4::Zero;
 
 		for (size_t i = 0; i < DebugTexts.size(); i++)
 		{
 			HDC ImageDc = GameEngineWindow::GetDoubleBufferImage()->GetImageDC();
-			TextOutA(ImageDc, TextOutStart.ix(), TextOutStart.iy(), DebugTexts[i].c_str(), DebugTexts[i].size());
+			TextOutA(ImageDc, TextOutStart.ix(), TextOutStart.iy(), DebugTexts[i].c_str(), static_cast<int>(DebugTexts[i].size()));
 			TextOutStart.y += 20.0f;
 		}
